+++ src/controllers/mainController.ts
@@ -32,7 +32,6 @@ import { QueryHistoryProvider } from '../queryHistory/queryHistoryProvider';
 import { QueryHistoryNode } from '../queryHistory/queryHistoryNode';
 import { DacFxService } from '../services/dacFxService';
 import { IConnectionInfo } from 'vscode-mssql';
-import { AzureFunctionProjectService } from '../azureFunction/azureFunctionProjectService';
 import { getConnectionString } from '../utils/connectionStringUtil';
 import { SchemaCompareService } from '../services/schemaCompareService';
 import { SqlTasksService } from '../services/sqlTasksService';
@@ -442,7 +441,6 @@ export default class MainController implements vscode.Disposable {
 				}
 			}));
 
-		const afService = new AzureFunctionProjectService(this.azureFunctionsService);
 		// Generate Azure Function command
 		this._context.subscriptions.push(vscode.commands.registerCommand(Constants.cmdCreateAzureFunction, async (node: TreeNodeInfo) => {
 			const database = ObjectExplorerUtils.getDatabaseName(node);
@@ -452,7 +450,7 @@ export default class MainController implements vscode.Disposable {
 				database,
 				node.connectionInfo.user,
 				node.connectionInfo.password);
-			await afService.createAzureFunction(connStr, node.metadata.schema, node.metadata.name);
+			await this.azureFunctionsService.createAzureFunction(connStr, node.metadata.schema, node.metadata.name);
 		}));
 	}
