+++ src/services/azureFunctionsService.ts
@@ -4,8 +4,13 @@
  * ------------------------------------------------------------------------------------------ */
 
 import SqlToolsServiceClient from '../languageservice/serviceclient';
+import * as vscode from 'vscode';
 import * as mssql from 'vscode-mssql';
 import * as azureFunctionsContracts from '../models/contracts/azureFunctions/azureFunctionsContracts';
+import * as azureFunctionUtils from '../azureFunction/azureFunctionUtils';
+import * as constants from '../constants/constants';
+import { generateQuotedFullName } from '../utils/utils';
+import * as LocalizedConstants from '../constants/localizedConstants';
 
 export const hostFileName: string = 'host.json';
 
@@ -54,5 +59,54 @@ export class AzureFunctionsService implements mssql.IAzureFunctionsService {
 		};
 
 		return this._client.sendRequest(azureFunctionsContracts.GetAzureFunctionsRequest.type, params);
+	}
+
+	public async createAzureFunction(connectionString: string, schema: string, table: string): Promise<void> {
+		const azureFunctionApi = await azureFunctionUtils.getAzureFunctionsExtensionApi();
+		if (!azureFunctionApi) {
+			return;
+		}
+		let projectFile = await azureFunctionUtils.getAzureFunctionProject();
+		if (!projectFile) {
+			vscode.window.showErrorMessage(LocalizedConstants.azureFunctionsProjectMustBeOpened);
+			return;
+		}
+
+		// because of an AF extension API issue, we have to get the newly created file by adding
+		// a watcher: https://github.com/microsoft/vscode-azurefunctions/issues/2908
+		const newFilePromise = azureFunctionUtils.waitForNewFunctionFile(projectFile);
+
+		// get function name from user
+		const functionName = await vscode.window.showInputBox({
+			title: constants.functionNameTitle,
+			value: table,
+			ignoreFocusOut: true
+		});
+		if (!functionName) {
+			return;
+		}
+
+		// create C# HttpTrigger
+		await azureFunctionApi.createFunction({
+			language: 'C#',
+			templateId: 'HttpTrigger',
+			functionName: functionName,
+			folderPath: projectFile
+		});
+
+		await azureFunctionUtils.addNugetReferenceToProjectFile(projectFile);
+		await azureFunctionUtils.addConnectionStringToConfig(connectionString, projectFile);
+		const functionFile = await newFilePromise;
+
+		let objectName = generateQuotedFullName(schema, table);
+		await this.addSqlBinding(
+			mssql.BindingType.input,
+			functionFile,
+			functionName,
+			objectName,
+			constants.sqlConnectionString
+		);
+
+		azureFunctionUtils.overwriteAzureFunctionMethodBody(functionFile);
 	}
 }
